+++ internal-architecture/payment-management-service/src/main/java/com/ms/payment_management_service/application/port/input/PaymentService.java
@@ -1,0 +1,8 @@
+package com.ms.payment_management_service.application.port.input;
+
+import com.ms.payment_management_service.domain.ElectricBillRequest;
+import com.ms.payment_management_service.domain.ElectricBillResponse;
+
+public interface PaymentService {
+    ElectricBillResponse getPaymentDetail(ElectricBillRequest electricBillRequest);
+}
