+++ internal-architecture/payment-management-service/src/main/java/com/ms/payment_management_service/application/port/output/ElectricBillRepository.java
@@ -1,0 +1,7 @@
+package com.ms.payment_management_service.application.port.output;
+
+import com.ms.payment_management_service.domain.ElectricBillResponse;
+
+public interface ElectricBillRepository {
+    ElectricBillResponse getBill(String providerId, String referenceNumber);
+}
