+++ internal-architecture/payment-management-service/src/main/java/com/ms/payment_management_service/domain/ElectricBillRequest.java
@@ -1,4 +1,4 @@
-package com.ms.payment_management_service;
+package com.ms.payment_management_service.domain;
 
 public record ElectricBillRequest(String idClient, String idService, String referenceNumber) {
 }
