+++ internal-architecture/payment-management-service/src/main/java/com/ms/payment_management_service/domain/ElectricBillResponse.java
@@ -1,4 +1,4 @@
-package com.ms.payment_management_service;
+package com.ms.payment_management_service.domain;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
