+++ internal-architecture/payment-management-service/src/main/java/com/ms/payment_management_service/infraestructure/repository/ElectricBillRepositoryImpl.java
@@ -1,0 +1,20 @@
+package com.ms.payment_management_service.infraestructure.repository;
+
+import com.ms.payment_management_service.application.port.output.ElectricBillRepository;
+import com.ms.payment_management_service.domain.ElectricBillResponse;
+import com.ms.payment_management_service.infraestructure.adapter.feign.ElectricBillClient;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class ElectricBillRepositoryImpl implements ElectricBillRepository {
+    private final ElectricBillClient electricBillClient;
+
+    public ElectricBillRepositoryImpl(ElectricBillClient electricBillClient) {
+        this.electricBillClient = electricBillClient;
+    }
+
+    @Override
+    public ElectricBillResponse getBill(String providerId, String referenceNumber) {
+        return electricBillClient.getBill(providerId,referenceNumber);
+    }
+}
